# docs/openapi.json
{
  "components": {
    "schemas": {
      "Body_post_upload_api_uploads_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_post_upload_api_uploads_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Analyzes radar diagnostic zipfiles with deterministic rules plus per-category LLM synthesis.",
    "title": "radar-analyst",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/config": {
      "get": {
        "description": "List the AI providers, with availability and model name.\n\nEach entry carries ``name``, ``label``, ``available`` and\n``model``; an unavailable provider that can explain itself also\ncarries ``reason``.",
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Config Api Config Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Config",
        "tags": [
          "config"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "description": "One job's state, phase, timestamps, and error.",
        "operationId": "get_job_status_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job Status Api Jobs  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Status",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/events": {
      "get": {
        "description": "Progress-event stream for one job.",
        "operationId": "stream_job_events_api_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Job Events",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/uploads": {
      "get": {
        "description": "Paginated uploads, newest first.",
        "operationId": "list_uploads_route_api_uploads_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Uploads Route Api Uploads Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Uploads Route",
        "tags": [
          "uploads"
        ]
      },
      "post": {
        "description": "Accept a radar zip, store it, enqueue analysis.",
        "operationId": "post_upload_api_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_upload_api_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Post Upload Api Uploads Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Upload",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/uploads/{upload_id}": {
      "delete": {
        "description": "Delete an upload, its blob, and cascaded rows.",
        "operationId": "delete_upload_route_api_uploads__upload_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Upload Route",
        "tags": [
          "uploads"
        ]
      },
      "get": {
        "description": "One upload's metadata.",
        "operationId": "get_upload_route_api_uploads__upload_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Upload Route Api Uploads  Upload Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Upload Route",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/uploads/{upload_id}/assess": {
      "post": {
        "description": "Assess the upload again from its stored archive.\n\nThe previous briefs and findings are removed first, so the\nassessment page never shows two results side by side; the new\nrun replaces the snapshot when it writes one. Refused with 409\nwhile an assessment of the upload is still running.",
        "operationId": "assess_again_route_api_uploads__upload_id__assess_post",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Assess Again Route Api Uploads  Upload Id  Assess Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assess Again Route",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/uploads/{upload_id}/assessment": {
      "get": {
        "description": "The upload's briefs plus the roll-up verdict over them.\n\nAn upload that hasn't been assessed yet returns an empty brief\nlist and a null verdict rather than a 404, because the console\npolls this while the job is still running.",
        "operationId": "get_assessment_route_api_uploads__upload_id__assessment_get",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Assessment Route Api Uploads  Upload Id  Assessment Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Assessment Route",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/uploads/{upload_id}/files": {
      "get": {
        "description": "Inventory of every entry in the upload's radar zip.\n\nReturns the persisted ``archive_files`` snapshot: populated\nby the orchestrator at analysis-time. An upload that exists\nbut hasn't been analysed yet returns ``items: []``.",
        "operationId": "list_files_route_api_uploads__upload_id__files_get",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Files Route Api Uploads  Upload Id  Files Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Files Route",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/uploads/{upload_id}/files/{archive_path}": {
      "get": {
        "description": "Stream a single entry out of the upload's radar zip.\n\nThe requested path must appear in the persisted inventory.\nAnything else (including ``..`` traversal attempts) is a\nplain 404: the inventory is the whitelist.",
        "operationId": "download_file_route_api_uploads__upload_id__files__archive_path__get",
        "parameters": [
          {
            "in": "path",
            "name": "archive_path",
            "required": true,
            "schema": {
              "title": "Archive Path",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download File Route",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/uploads/{upload_id}/snapshot": {
      "get": {
        "description": "The upload's parsed snapshot document.",
        "operationId": "get_snapshot_route_api_uploads__upload_id__snapshot_get",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Snapshot Route Api Uploads  Upload Id  Snapshot Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Snapshot Route",
        "tags": [
          "uploads"
        ]
      }
    },
    "/healthz": {
      "get": {
        "description": "Liveness probe: static ok.",
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Healthz Healthz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz"
      }
    },
    "/readyz": {
      "get": {
        "description": "Readiness probe: ok only once the DB pool is attached.",
        "operationId": "readyz_readyz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Readyz Readyz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readyz"
      }
    }
  }
}